Fix Constants (now in each file) and MainConstants (for all project)
