--- EarlierBird/Constants/Constants.swift
@@ -1,49 +1,0 @@
-//
-//  Constants.swift
-//  EarlierBird
-//
-//  Created by Konstantin Gracheff on 27.02.2022.
-//
-
-import UIKit
-
-enum Constants {
-    enum ProgressView {
-        static let calendarIconDistanceToTopEdge: CGFloat = 10
-        static let calendarIconHeightWidth: CGFloat = 60
-        static let textDistanceToTopEdge: CGFloat = 20
-        static let progressViewHeight: CGFloat = 20
-        static let progressViewDistanceToTopEdge: CGFloat = 15
-        static let progressIconsHeightWidth: CGFloat = 35
-    }
-    
-    enum TodayAlarmView {
-        static let digitalDialFaceIconToTopEdge: CGFloat = 2
-        static let digitalDialFaceIconHeight: CGFloat = 90
-    }
-    
-    enum MainViewController {
-        static let timeLabelStackViewSpacing: CGFloat = -13
-        static let timeLabelStackViewWidth: CGFloat = 130
-        static let goalImageHeightMultiplier: CGFloat = 0.9
-    }
-    
-    enum SettingsController {
-        static let spacingHorizontalStackView: CGFloat = 10
-        static let spacingVerticalStackView: CGFloat = 5
-        static let spacingGeneralStackView: CGFloat = 10
-        static let notificationButtonHeight: CGFloat = 45
-    }
-    
-    enum ContraintsSize {
-        static let defaultDisctanceToSideEdge: CGFloat = 20
-        static let defaultDistanceToTopEdge: CGFloat = 10
-        static let defaultDistanceToTopSafeAreaEdge: CGFloat = 0
-        static let defaultDistanceBetweenViews: CGFloat = 20
-        static let defaultViewHeight: CGFloat = 130
-    }
-    
-    enum CornerRadius {
-        static let defaultCornerRadius: CGFloat = 10
-    }
-}
