+++ EarlierBird/Constants/MainConstants.swift
@@ -1,0 +1,23 @@
+//
+//  Constants.swift
+//  EarlierBird
+//
+//  Created by Konstantin Gracheff on 27.02.2022.
+//
+
+import UIKit
+
+enum MainConstants {
+    
+    enum ContraintsSize {
+        static let defaultDisctanceToSideEdge: CGFloat = 20
+        static let defaultDistanceToTopEdge: CGFloat = 10
+        static let defaultDistanceToTopSafeAreaEdge: CGFloat = 0
+        static let defaultDistanceBetweenViews: CGFloat = 20
+        static let defaultViewHeight: CGFloat = 130
+    }
+    
+    enum CornerRadius {
+        static let defaultCornerRadius: CGFloat = 10
+    }
+}
